pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "brain"
version = "0.1.0"
description = "SSE Server & AI Processing Engine"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Andre Rogers"}
]
dependencies = [
    "anthropic>=0.52.2",
    "exa-py>=1.14.0",
    "fastapi",
    "langchain-mcp-adapters>=0.1.7",
    "logfire",
    "mcp>=1.8.0",
    "psutil>=7.0.0",
    "pydantic-ai>=0.0.14",
    "setuptools>=80.4.0",
    "uvicorn",
]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "I"]
